requirements.txt
fastapi
uvicorn
pydantic
Pillow
azure-ai-formrecognizer
python-multipart
boto3
python-dotenv
sqlalchemy
pymysql
